web/fix: resolved merge conflicts
