+++ packages/web/app/_components/tasks/tasks-page.tsx
@@ -2,7 +2,6 @@
 
 import { FC, useState } from 'react';
 import { Card, Group, Stack } from '@mantine/core';
-import { Task } from '@prisma/client';
 import { useProjectsTabs } from '@/hooks/use-projects-tabs';
 import { Project } from '@prisma/client';
 import { ProjectsTabs } from '../projects-tabs';
@@ -15,16 +14,9 @@ export const TasksPage: FC = () => {
     projectId,
     projectName,
     tabs,
-    selectedTask,
     popoverOpen,
     handleChange,
-    onEdit,
-    onDelete,
-    onReorder,
-    onSubmit,
     setPopoverOpen,
-    handleTaskSelect,
-    handlePopoverAfterClose,
   } = useTasksPage();
 
   const renderProject = ({ id, name }: Project) => (
@@ -87,37 +79,14 @@ function useTasksPage() {
   const { projects, selectedProject, tabs, handleChange } =
     useProjectsTabs(true);
   const { id: projectId, name: projectName } = selectedProject ?? {};
-  const [
-    { todos, done, selectedTask },
-    { onSelectTask, onEdit, onDelete, onReorder, onSubmit },
-  ] = useTasks(projectId!, {
-    onSubmitSuccess: () => {
-      onSelectTask(null);
-      setPopoverOpen(false);
-    },
-  });
-
-  const handleTaskSelect = (task: Task) => {
-    onSelectTask(task);
-    setPopoverOpen(true);
-  };
-
-  const handlePopoverAfterClose = () => onSelectTask(null);
 
   return {
     projects: projects?.data ?? [],
     projectId,
     projectName,
     tabs,
-    selectedTask,
     popoverOpen,
-    handleChange,
-    onEdit,
-    onDelete,
-    onReorder,
-    onSubmit,
     setPopoverOpen,
-    handleTaskSelect,
-    handlePopoverAfterClose,
+    handleChange,
   };
 }
